Resolves issue of not being able to pull personID from parent by making the insertParent method return the personID field.  Modifies layout of player Details page, and add Parent Details page.

+++ app/src/main/java/android/reserver/capstone_robertklare/Entities/Person.java
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey;
 public class Person {
 
     @PrimaryKey(autoGenerate = true)
-    int personID;
+    long personID;
     String firstName;
     String lastname;
     int teamID;
@@ -18,7 +18,7 @@ public class Person {
         this.teamID = teamID;
     }
 
-    public int getPersonID() {
+    public long getPersonID() {
         return personID;
     }
